requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
openai>=1.0.0 